# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nova Forum Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Получает список тем",
                "description": "Возвращает список популярных и последних тем.",
                "operationId": "f90ade7e8faeb644e4910b74e974661a",
                "responses": {
                    "200": {
                        "description": "Успешный ответ с данными тем",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "properties": {
                                                "latest": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "uuid": {
                                                                "type": "string"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "preview": {
                                                                "type": "string"
                                                            },
                                                            "user_id": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "popular": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "uuid": {
                                                                "type": "string"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "created_at": {
                                                                "type": "string"
                                                            },
                                                            "updated_at": {
                                                                "type": "string"
                                                            },
                                                            "preview": {
                                                                "type": "string"
                                                            },
                                                            "user_id": {
                                                                "type": "integer"
                                                            },
                                                            "comments_count": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 400
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Неверный запрос."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запрашиваемый ресурс не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 404
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Запрашиваемый ресурс не найден."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Произошла непредвиденная ошибка: [сообщение об ошибке]."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/search": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Получает список тем",
                "description": "Возвращает список популярных тем по ключевым фразам в описании и названии.",
                "operationId": "9579aca428084dea5083598dc064bd7e",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Ключевая фраза для поиска тем",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ с данными тем",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "properties": {
                                                "latest": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "uuid": {
                                                                "type": "string"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "preview": {
                                                                "type": "string"
                                                            },
                                                            "user_id": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "popular": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "uuid": {
                                                                "type": "string"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "preview": {
                                                                "type": "string"
                                                            },
                                                            "user_id": {
                                                                "type": "integer"
                                                            },
                                                            "comments_count": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 400
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Неверный запрос."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запрашиваемый ресурс не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 404
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Запрашиваемый ресурс не найден."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Произошла непредвиденная ошибка: [сообщение об ошибке]."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}